// src/app/shared/types/movieTypes.ts
export type SearchResult = {
  title: string;
  poster: string;
  year: string;
  imdbID: string;
};

export type Movie = {
  title: string;
  poster: string;
  rated: string;
  released: string;
  runtime: string;
  genre: string;
  director: string;
  writer: string;
  actors: string;
  plot: string;
  language: string;
  country: string;
  awards: string;
  ratings: string;
  metascore: string;
  imdbRating: string;
  imdbVotes: string;
  imdbID: string;
  type: string;
  dvd: string;
  boxOffice: string;
  production: string;
  website: string;
  response: string;
};